+++ advent-of-code/2024/AoC2024/13-claw-contraption/ClawContraption.PartTwo.cs
@@ -2,15 +2,15 @@ namespace AoC2024;
 
 public partial class ClawContraption
 {
-    public static int PartTwo(string filePath) =>
+    public static long PartTwo(string filePath) =>
         Parse(filePath)
             .AsParallel()
             .Select(config => new MachineConfig(config.A, config.B, AddOffset(config.Prize)))
             .Select(GetMinimumCost)
-            .Where(cost => cost != int.MaxValue)
+            .Where(cost => cost != long.MaxValue)
             .Sum();
     
-    private static readonly ulong offset = 10_000_000_000_000;
+    private static readonly long offset = 10_000_000_000_000;
 
     private static Vector AddOffset(Vector v) => new(v.X + offset, v.Y + offset);
 }
